requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
sqlalchemy==2.0.30
alembic==1.13.1
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
pydantic==2.7.1
pydantic-settings==2.2.1
httpx==0.27.0
aiofiles==23.2.1
python-dotenv==1.0.1

# HMR2.0 / 4D-Humans inference
torch==2.3.0
torchvision==0.18.0
# 4dhuman package installed separately — see README
smplx==0.1.28
trimesh==4.3.2
numpy==1.26.4
opencv-python-headless==4.9.0.80
